pyproject.toml
[project]
name = "image_recog_ml"
version = "0.1.0"
authors = [{ name = "Phe Pato", email = "pheprogrammer@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12"

[tool.poetry.dependencies]
toml = ">=0.10.2,<0.11.0"
pydantic = ">=2.11.4,<3.0.0"
numpy = ">=2.2.5,<3.0.0"
opencv-python = ">=4.11.0.86,<5.0.0.0"
duckdi = ">=0.1.13,<0.2"
simple-duck-ml = "^0.1.2"
fastapi = {extras = ["standart"], version = "^0.119.0"}
uvicorn = "^0.38.0"
jinja2 = "^3.1.6"
pillow = "^12.0.0"
pyrasite = "^2.0"

[tool.poetry.group.dev.dependencies]
isort = ">=6.0.1,<7.0.0"
black = ">=25.1.0,<26.0.0"
mypy = ">=1.15.0,<2.0.0"
bandit = ">=1.8.3,<2.0.0"
flake8 = ">=7.2.0,<8.0.0"
autopep8 = ">=2.3.2,<3.0.0"
pytest = ">=8.3.5,<9.0.0"

[tool.pytest.ini_options]
pythonpath = ["tests"]
log_cli = true

[tool.poetry]
packages = [
  {include = "*", from = "src"},
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
